requirements.txt
gradio==4.44.1
sqlalchemy
python-dotenv
